+++ crates/minerva/src/attribute_store/materialize_curr_ptr.rs
@@ -6,7 +6,7 @@ use crate::error::Error;
 
 pub struct MaterializeCurrPtrResult {
     pub record_count: u64,
-    pub materialized_view: bool,
+    pub updated_curr_data: String,
     pub attribute_store_name: String,
     pub attribute_store_id: i32,
 }
@@ -89,31 +89,44 @@ pub async fn materialize_curr_ptr_by_name<T: GenericClient + Send + Sync>(
 
     let record_count: i32 = row.get(0);
 
-    let materialized_view = if has_materialized_view(client, attribute_store_name).await? {
-        refresh_materialized_view(client, attribute_store_name).await?;
-
-        true
-    } else {
-        false
+    let updated_curr_data = match curr_data_storage_method(client, attribute_store_name).await? {
+        CurrDataStorageMethod::View => {
+            // Nothing to do for a view
+            "view".to_string()
+        }
+        CurrDataStorageMethod::Table => {
+            update_curr_table(client, attribute_store_name).await?;
+            "table".to_string()
+        }
+        CurrDataStorageMethod::MaterializedView => {
+            refresh_materialized_view(client, attribute_store_name).await?;
+            "materialized view".to_string()
+        }
     };
 
     Ok(MaterializeCurrPtrResult {
         record_count: record_count.try_into().unwrap(),
-        materialized_view,
+        updated_curr_data,
         attribute_store_name: attribute_store_name.to_string(),
         attribute_store_id,
     })
 }
 
-pub async fn has_materialized_view<T: GenericClient + Send + Sync>(
+pub enum CurrDataStorageMethod {
+    View,
+    MaterializedView,
+    Table,
+}
+
+pub async fn curr_data_storage_method<T: GenericClient + Send + Sync>(
     client: &T,
     attribute_store_name: &str,
-) -> Result<bool, MaterializeCurrPtrError> {
+) -> Result<CurrDataStorageMethod, MaterializeCurrPtrError> {
     let query = concat!(
-        "SELECT relname ",
+        "SELECT relkind ",
         "FROM pg_class c ",
         "JOIN pg_namespace ns ON ns.oid = c.relnamespace ",
-        "WHERE ns.nspname = 'attribute' AND c.relname = $1 AND c.relkind = 'm'"
+        "WHERE ns.nspname = 'attribute' AND c.relname = $1"
     );
 
     let rows = client
@@ -126,10 +139,70 @@ pub async fn has_materialized_view<T: GenericClient + Send + Sync>(
         })?;
 
     if rows.is_empty() {
-        Ok(false)
+        Err(MaterializeCurrPtrError::Unexpected(
+            "Could not determine current data storage method: No meta data found".to_string(),
+        ))
     } else {
-        Ok(true)
+        let row = rows.first().unwrap();
+
+        let kind: i8 = row.get(0);
+        let kind_char = kind as u8 as char;
+
+        match kind_char {
+            'm' => Ok(CurrDataStorageMethod::MaterializedView),
+            'r' => Ok(CurrDataStorageMethod::Table),
+            'v' => Ok(CurrDataStorageMethod::View),
+            _ => Err(MaterializeCurrPtrError::Unexpected(format!(
+                "Unexpected relation type: '{}'",
+                kind_char
+            ))),
+        }
     }
+}
+
+pub async fn update_curr_table<T: GenericClient + Send + Sync>(
+    client: &T,
+    attribute_store_name: &str,
+) -> Result<u64, MaterializeCurrPtrError> {
+    let query = concat!(
+        "SELECT attribute.name ",
+        "FROM attribute_directory.attribute ",
+        "JOIN attribute_directory.attribute_store ast ON ast.id = attribute.attribute_store_id ",
+        "WHERE ast::text = $1",
+    );
+
+    let rows = client
+        .query(query, &[&attribute_store_name])
+        .await
+        .map_err(|e| {
+            MaterializeCurrPtrError::Unexpected(format!("Could not query attributes: {e}"))
+        })?;
+
+    let mut columns: Vec<String> = vec![
+        "id",
+        "first_appearance",
+        "modified",
+        "hash",
+        "entity_id",
+        "end",
+    ]
+    .into_iter()
+    .map(String::from)
+    .collect();
+    let mut attribute_names: Vec<String> = rows.iter().map(|row| row.get(0)).collect();
+    columns.append(&mut attribute_names);
+
+    let cols_part = columns.join(",");
+
+    let query = format!("INSERT INTO attribute.\"{}\"({}) SELECT {} FROM attribute_history.\"{}\" history JOIN attribute_history.\"{}_curr_ptr\" curr_ptr ON history.id = curr_ptr.id", attribute_store_name, cols_part, cols_part, attribute_store_name, attribute_store_name);
+
+    let record_count = client.execute(&query, &[]).await.map_err(|e| {
+        MaterializeCurrPtrError::Unexpected(format!(
+            "Could not load data into curr-data table: {e}"
+        ))
+    })?;
+
+    Ok(record_count)
 }
 
 pub async fn refresh_materialized_view<T: GenericClient + Send + Sync>(